requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
reportlab>=3.6.12
Pillow>=9.4.0
python-multipart>=0.0.6
aiosmtplib>=3.0.0
email-validator>=2.1.0

# Gmail API for draft management
